Agents/Handoffs/pyproject.toml
[project]
name = "handoffs"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai-agents>=0.1.0",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
]
